splits integration tests even further and prepares for separate directories
